Move AggregateFunctionRewriter to plugin toolkit
This is to make it reusable between connectors based on base-jdbc and
the Pinot connector.
