+++ lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/expression/AggregateFunctionPatterns.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.jdbc.expression;
+package io.trino.plugin.base.expression;
 
 import com.google.common.collect.ImmutableList;
 import io.trino.matching.Captures;
