+++ plugin/trino-base-jdbc/src/main/java/io/trino/plugin/jdbc/expression/ImplementAvgFloatingPoint.java
@@ -16,6 +16,7 @@ package io.trino.plugin.jdbc.expression;
 import io.trino.matching.Capture;
 import io.trino.matching.Captures;
 import io.trino.matching.Pattern;
+import io.trino.plugin.base.expression.AggregateFunctionRule;
 import io.trino.plugin.jdbc.JdbcColumnHandle;
 import io.trino.plugin.jdbc.JdbcExpression;
 import io.trino.spi.connector.AggregateFunction;
@@ -25,11 +26,11 @@ import java.util.Optional;
 
 import static com.google.common.base.Verify.verify;
 import static io.trino.matching.Capture.newCapture;
-import static io.trino.plugin.jdbc.expression.AggregateFunctionPatterns.basicAggregation;
-import static io.trino.plugin.jdbc.expression.AggregateFunctionPatterns.expressionType;
-import static io.trino.plugin.jdbc.expression.AggregateFunctionPatterns.functionName;
-import static io.trino.plugin.jdbc.expression.AggregateFunctionPatterns.singleInput;
-import static io.trino.plugin.jdbc.expression.AggregateFunctionPatterns.variable;
+import static io.trino.plugin.base.expression.AggregateFunctionPatterns.basicAggregation;
+import static io.trino.plugin.base.expression.AggregateFunctionPatterns.expressionType;
+import static io.trino.plugin.base.expression.AggregateFunctionPatterns.functionName;
+import static io.trino.plugin.base.expression.AggregateFunctionPatterns.singleInput;
+import static io.trino.plugin.base.expression.AggregateFunctionPatterns.variable;
 import static io.trino.spi.type.DoubleType.DOUBLE;
 import static io.trino.spi.type.RealType.REAL;
 import static java.lang.String.format;
@@ -38,7 +39,7 @@ import static java.lang.String.format;
  * Implements {@code avg(float)}
  */
 public class ImplementAvgFloatingPoint
-        implements AggregateFunctionRule
+        implements AggregateFunctionRule<JdbcExpression>
 {
     private static final Capture<Variable> INPUT = newCapture();
