+++ plugin/trino-clickhouse/pom.xml
@@ -27,6 +27,12 @@
             <groupId>io.trino</groupId>
             <artifactId>trino-matching</artifactId>
         </dependency>
+
+        <dependency>
+            <groupId>io.trino</groupId>
+            <artifactId>trino-plugin-toolkit</artifactId>
+        </dependency>
+
         <dependency>
             <groupId>io.airlift</groupId>
             <artifactId>configuration</artifactId>
